# api/docs/swagger.yaml
basePath: /
definitions:
  apperror.Code:
    enum:
    - INVALID_CREDENTIALS
    - USER_NOT_FOUND
    - USER_INACTIVE
    - INVALID_TOKEN
    - VALIDATION_ERROR
    - INTERNAL_ERROR
    - INVALID_JD_INPUT
    - JD_TOO_SHORT
    - JD_TOO_LONG
    - EXTRACTION_FAILED
    - INVALID_EXTRACTION_OUTPUT
    type: string
    x-enum-varnames:
    - CodeInvalidCredentials
    - CodeAccountNotFound
    - CodeAccountLocked
    - CodeInvalidToken
    - CodeValidation
    - CodeInternal
    - CodeInvalidJDInput
    - CodeJDTooShort
    - CodeJDTooLong
    - CodeExtractionFailed
    - CodeInvalidExtractionOutput
  auth.LoginRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      password:
        example: SuperSecret123
        type: string
    required:
    - email
    - password
    type: object
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      message:
        type: string
    type: object
info:
  contact: {}
  description: Existing backend foundation endpoints. Health is liveness only; login
    returns a normalized envelope.
  title: AI Interview Practice API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns an access token in data and sets the refresh cookie scoped
        to /auth/refresh. A refresh endpoint is not currently registered.
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access token
          headers:
            Set-Cookie:
              description: Refresh token cookie; Path=/auth/refresh
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Account locked
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Log in with email and password
      tags:
      - auth
  /health:
    get:
      description: Does not check database or LLM readiness.
      produces:
      - application/json
      responses:
        "200":
          description: 'status: ok'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check process liveness
      tags:
      - health
swagger: "2.0"